Extract method to services
