+++ server/src/services/report_status.py
@@ -1,7 +1,9 @@
+from collections import defaultdict
 import json
 import logging
 import threading
 from datetime import UTC, datetime
+from typing import Dict
 
 import requests
 
@@ -233,3 +235,26 @@ def update_report_config(slug: str, updated_config: ReportConfigUpdate) -> dict:
 
     invalidate_report_cache(slug)
     return _report_status[slug]
+
+
+def add_analysis_data(report: Report):
+    if (report.status == ReportStatus.READY):
+        new_report_dict = report.__dict__.copy()
+        report_path = settings.REPORT_DIR / report.slug / "hierarchical_result.json"
+        with open(report_path) as f:
+            report_result = json.load(f)
+            new_report_dict["analysis"] = {
+                "comment_num": report_result["comment_num"],
+                "arguments_num": len(report_result["arguments"]),
+                "cluster_num": get_cluster_num(report_result)
+            }
+        return new_report_dict
+    else:
+        return report
+
+def get_cluster_num(result: dict) -> Dict[int, int]:
+    array = [c["level"] for c in result["clusters"]]
+    acc = defaultdict(int)
+    for num in array:
+        acc[num] += 1
+    return dict(acc)[2]
